tweak(pages): embed SVG icons in DOM
Using a custom element `SvgIcon`.
Mainly to avoid using filters in the CSS to color the icons, which was interfering with other elements, i.e., the outline color for the "copy link" and "copy code snippet" buttons. Now we can simply use the `color` property in CSS.
The icons in the dump tree still use .svg through CSS backgrounds.