fix(pages): fix TS using types from NodeJS
Installing webpack added node_modules/@types/node which TS used by default.

This was causing issues with `setTimeout` because in NodeJS it returns `NodeJS.Timeout` while in the browser it returns `number`.

And webpack's ts-loader didn't seem to be using our tsconfig.json as it was still trying to use `NodeJS.Timeout`. Now the tsconfig is explicitly defined.