tweak(fmt): support GTA4 dumps
Mainly the Name type now supports literal strings in the JSON instead of just hashes (if the string starts with `0x` it is treated as a hash; otherwise, as a literal string). As such, properties that provided literal string names when available have been removed (`ParStructure.NameStr`, `ParStructure.MemberNames` and `ParEnum.ValueNames`) in favor of storing the literal string name directly in the corresponding name property.

And `ParDump.Build` is now a string as it will contain the full version, not just a single number.